refactor(server): 升级协议指令类型与内存池实现
1. 统一将业务指令ID从uint16升级为uint32，适配更大的指令空间
2. 重构内存池实现，替换为分级切片池提升内存复用效率
3. 新增数据长度字段到工作任务，完善任务元信息
4. 调整依赖包位置，将bytebufferpool移至间接依赖

+++ server/cmd/main.go
@@ -63,10 +63,10 @@ func main() {
 	router := handler.NewRouter()
 
 	// 注册测试用 Handler
-	router.Register(uint16(protocol.CmdCalculate), &handler.CalculateHandler{})
-	router.Register(uint16(protocol.CmdSmall), &handler.SmallHandler{})
-	router.Register(uint16(protocol.CmdMedium), &handler.MediumHandler{})
-	router.Register(uint16(protocol.CmdLarge), &handler.LargeHandler{})
+	router.Register(protocol.CmdCalculate, &handler.CalculateHandler{})
+	router.Register(protocol.CmdSmall, &handler.SmallHandler{})
+	router.Register(protocol.CmdMedium, &handler.MediumHandler{})
+	router.Register(protocol.CmdLarge, &handler.LargeHandler{})
 
 	// 初始化业务协程池
 	workerPool := pool.InitWorkerPool(config.Global.Server.WorkerPoolSize,
